pyproject.toml
[tool.poetry]
name = "pasta-la-pepito"
version = "0.1.0"
description = ""
authors = ["Oleh Sofronov <olehsofronov@gmail.com>, Vladislav Gryadko <vgryadko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}
djoser = "^2.2.2"
drf-spectacular = "^0.27.1"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
python-dotenv = "^1.0.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
